@@ -3,42 +3,43 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "nodemon --watch ./src -e ts --exec 'pnpm ts-node --files ./src/index.ts'",
-    "build": "prisma generate && tsc",
+    "dev": "vite dev",
+    "build": "prisma generate && vite build",
     "start": "node dist/index.js",
     "lint": "eslint .",
-    "cypress": "cypress open",
-    "cypress:headless": "cypress run",
-    "e2e": "pnpm build && start-test 'pnpm start' http://0.0.0.0:1234/ 'pnpm cypress'",
-    "e2e:headless": "pnpm build && start-test 'pnpm start' http://0.0.0.0:1234/ 'pnpm cypress:headless'"
+    "test": "vitest",
+    "test:headless": "prisma generate && vitest run",
+    "coverage": "vitest run --coverage"
   },
   "prisma": {
     "schema": "src/database/schema.prisma"
   },
   "dependencies": {
-    "@prisma/client": "^3.14.0",
+    "@opensom/config": "workspace:*",
+    "@prisma/client": "^3.15.1",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/jsonwebtoken": "^8.5.8",
-    "@types/node": "^17.0.32",
+    "@types/multer": "^1.4.7",
+    "@types/node": "^17.0.41",
     "@types/nodemailer": "^6.4.4",
     "cors": "^2.8.5",
     "date-fns": "^2.28.0",
-    "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-async-errors": "^3.1.1",
     "jsonwebtoken": "^8.5.1",
+    "multer": "1.4.5-lts.1",
     "nodemailer": "^6.7.5",
-    "prisma": "^3.14.0",
-    "typescript": "^4.6.4",
-    "@opensom/config": "workspace:*"
+    "prisma": "^3.15.1",
+    "typescript": "^4.7.3",
+    "vite": "^2.9.10",
+    "vite-plugin-node": "^1.0.0"
   },
   "devDependencies": {
-    "@faker-js/faker": "^6.3.1",
-    "cypress": "^9.6.1",
-    "eslint-plugin-cypress": "^2.12.1",
-    "nodemon": "^2.0.16",
-    "start-server-and-test": "^1.14.0",
-    "ts-node": "^10.7.0"
+    "@faker-js/faker": "^7.2.0",
+    "@types/supertest": "^2.0.12",
+    "c8": "^7.11.3",
+    "supertest": "^6.2.3",
+    "vitest": "^0.14.1"
   }
 }
