@@ -9,26 +9,26 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@date-io/date-fns": "^2.13.2",
+    "@date-io/date-fns": "^2.14.0",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.7.0",
-    "@mui/x-date-pickers": "5.0.0-alpha.2",
-    "@types/node": "^17.0.32",
-    "@types/react": "^18.0.9",
+    "@mui/icons-material": "^5.8.3",
+    "@mui/material": "^5.8.3",
+    "@mui/x-date-pickers": "5.0.0-alpha.5",
+    "@opensom/config": "workspace:*",
+    "@types/node": "^17.0.41",
+    "@types/react": "^18.0.12",
     "date-fns": "^2.28.0",
-    "easyqrcodejs": "^4.4.10",
+    "easyqrcodejs": "^4.4.12",
     "next": "^12.1.6",
     "nookies": "^2.5.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "sass": "^1.51.0",
+    "sass": "^1.52.3",
     "swr": "^1.3.0",
-    "typescript": "^4.6.4",
-    "@opensom/config": "workspace:*"
+    "typescript": "^4.7.3"
   },
   "devDependencies": {
-    "concurrently": "^7.1.0"
+    "concurrently": "^7.2.1"
   }
 }
