@@ -11,19 +11,19 @@
   "dependencies": {
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.7.0",
+    "@mui/icons-material": "^5.8.3",
+    "@mui/material": "^5.8.3",
     "@opensom/config": "workspace:*",
-    "@types/node": "^17.0.32",
-    "@types/react": "^18.0.9",
+    "@types/node": "^17.0.41",
+    "@types/react": "^18.0.12",
     "date-fns": "^2.28.0",
     "next": "^12.1.6",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "sass": "^1.51.0",
-    "typescript": "^4.6.4"
+    "sass": "^1.52.3",
+    "typescript": "^4.7.3"
   },
   "devDependencies": {
-    "concurrently": "^7.1.0"
+    "concurrently": "^7.2.1"
   }
 }
