@@ -1,4 +1,4 @@
-module.exports = (() => {
+export default (() => {
   return {
     COMPANY: {
       name: "OpenSOM",
@@ -15,5 +15,6 @@ module.exports = (() => {
     },
     ATTENDANTS: ["Dan Cordova", "Larissa Brain", "Tom Coffee"],
     SERVICE_PLACES: ["Balcão", "Telefone"],
+    VOLTAGES_OPTIONS: ["Não se aplica", "Não identificável", "127V", "220V", "380V"], // Options that not starts with numbers will not be printed in PDF. They are for internal use only.
   };
 })();
