@@ -1,7 +1,9 @@
-import User from "./src/types/user";
+import config from "./config";
+
+type User = Omit<typeof config.users[0], "password">;
 
 declare global {
-  declare namespace Express {
+  namespace Express {
     export interface Request {
       user?: User;
     }
