@@ -1,10 +1,11 @@
-import "dotenv/config";
-
 import server from "./app";
 
-// Constants
-const port = process.env.PORT || 1234;
+const port = import.meta.env.PORT || 1234;
 console.log(`Server listening on port ${port}`);
 
 // Start server
-server.listen(port);
+if (import.meta.env.PROD) {
+  server.listen(port);
+}
+
+export const viteNodeApp = server;
