@@ -1,12 +1,10 @@
 import type { NextFunction, Request, Response } from "express";
 import jwt from "jsonwebtoken";
 
-import type User from "../types/user";
-
 const protect = async (req: Request, res: Response, next: NextFunction) => {
   const token = req.headers.authorization;
 
-  if (process.env.JWT_SECRET == undefined) {
+  if (import.meta.env.JWT_SECRET == undefined) {
     throw new Error("JWT_SECRET env variable not set");
   }
 
@@ -15,8 +13,8 @@ const protect = async (req: Request, res: Response, next: NextFunction) => {
   }
 
   try {
-    const decoded = jwt.verify(token, process.env.JWT_SECRET);
-    req.user = decoded as User;
+    const decoded = jwt.verify(token, import.meta.env.JWT_SECRET);
+    req.user = decoded as Request["user"];
     return next();
   } catch (error) {
     throw new Error("Unauthorized: Invalid token");
