@@ -1,7 +1,6 @@
+import config from "@config";
 import { Router } from "express";
 import jwt from "jsonwebtoken";
-
-import config from "../../config";
 
 const user = config.users[0];
 
@@ -10,7 +9,7 @@ const router = Router();
 router.post("", async (req, res) => {
   const { password } = req.body;
 
-  if (process.env.JWT_SECRET == undefined) {
+  if (import.meta.env.JWT_SECRET == undefined) {
     throw new Error("JWT_SECRET env variable not set");
   }
 
@@ -20,7 +19,7 @@ router.post("", async (req, res) => {
 
   const userWithoutPassword = { ...user, password: undefined };
 
-  const token = jwt.sign(userWithoutPassword, process.env.JWT_SECRET, {
+  const token = jwt.sign(userWithoutPassword, import.meta.env.JWT_SECRET, {
     expiresIn: 86400 * 7, // expires in 7 days
   });
 
