@@ -1,8 +1,6 @@
 import { Request, Response, Router } from "express";
 
-import { create, deleteOne, getAll, update } from "../../services/client";
-import { filterNumber } from "../../utils/filters";
-import { parseClient } from "../../utils/parsers";
+import { create, deleteOne, getAll, update } from "@/modules/client/controller";
 
 const router = Router();
 
@@ -12,30 +10,16 @@ router.get("", async (req: Request, res: Response) => {
 
 router.post("", async (req: Request, res: Response) => {
   const { client, serviceOrder } = req.body;
-  const newClient = await create(parseClient(client, serviceOrder));
-
-  return res.status(200).json(newClient);
+  return res.status(200).json(await create(client, serviceOrder));
 });
 
 router.put("", async (req: Request, res: Response) => {
-  const { client, serviceOrder } = req.body;
-  const id = filterNumber(req.body.clientID);
-
-  if (id === null) {
-    throw new Error("Invalid data: clientID");
-  }
-
-  return res.status(200).json(await update(id, parseClient(client, serviceOrder)));
+  const { clientId, client, serviceOrder } = req.body;
+  return res.status(200).json(await update(clientId, client, serviceOrder));
 });
 
 router.delete("", async (req: Request, res: Response) => {
-  const id = filterNumber(req.body.id);
-  if (id === null) {
-    throw new Error("Invalid data: id");
-  }
-
-  await deleteOne(id);
-  return res.status(200).json({ deletedID: id });
+  return res.status(200).json(await deleteOne(req.body.id));
 });
 
 export default router;
