@@ -1,13 +1,12 @@
 import { Request, Response, Router } from "express";
+import multer from "multer";
 
-import sendMail from "../../services/sendMail";
+import { sendMail } from "@/modules/mail/controller";
 
 const router = Router();
 
-router.post("", async (req: Request, res: Response) => {
-  const testMessageUrl = await sendMail(req.body.to, req.body.text);
-
-  return res.status(200).json({ status: "success", testMessageUrl });
+router.post("", multer().single("attachment"), async (req: Request, res: Response) => {
+  return res.status(200).json(await sendMail(req.body.to, req.body.text, req.file));
 });
 
 export default router;
