@@ -3,12 +3,12 @@ import useSWR from "swr";
 import { TServiceOrderWithClient } from "@/types/serviceOrder";
 
 export default function useServiceOrder(id: string) {
-  const { data, error, mutate } = useSWR(`/api/private/serviceOrders?id=${id}`);
+  const { data, error, mutate } = useSWR(`/api/private/serviceOrders/${id}`);
 
   return {
     serviceOrder: data as TServiceOrderWithClient | undefined,
     isLoading: !error && !data,
-    isError: error,
     mutate,
+    error,
   };
 }
