@@ -8,6 +8,7 @@ interface Req {
   url: string;
   method: Method;
   body?: unknown;
+  notJson?: boolean;
 }
 
 export default async function request(req: Req, ctx?: Pick<NextPageContext, "req">) {
@@ -18,7 +19,7 @@ export default async function request(req: Req, ctx?: Pick<NextPageContext, "req
 
   // Set headers
   const headers: HeadersInit = {};
-  if (body) {
+  if (body && !req.notJson) {
     headers["Content-Type"] = "application/json";
   }
   const token = parseCookies(ctx).token;
@@ -29,18 +30,18 @@ export default async function request(req: Req, ctx?: Pick<NextPageContext, "req
     res = await fetch(`${config.BACKEND_URL}${url}`, {
       method,
       headers,
-      body: body ? JSON.stringify(body) : undefined,
+      body: body ? (req.notJson ? (body as BodyInit) : JSON.stringify(body)) : undefined,
     });
     json = await res.json();
   } catch (e) {
     console.log(e);
   }
 
-  if (res?.status === 200) {
-    return { response: json, status: res.status };
-  } else {
+  if (res?.status !== 200) {
     return { error: getHumanReadableError(res), status: res?.status };
   }
+
+  return { response: json, status: res.status };
 }
 
 const getHumanReadableError = (res?: Response) => {
@@ -54,9 +55,15 @@ const getHumanReadableError = (res?: Response) => {
   else return "Erro desconhecido";
 };
 
-export const fetcher = (url: string) => {
-  return fetch(`${config.BACKEND_URL}${url}`, {
+export const fetcher = async (url: string) => {
+  const res = await fetch(`${config.BACKEND_URL}${url}`, {
     method: "GET",
     headers: { authorization: parseCookies().token },
-  }).then((r) => r.json());
+  });
+
+  if (!res.ok || res.status !== 200) {
+    throw getHumanReadableError(res);
+  }
+
+  return res.json();
 };
