@@ -1,9 +1,7 @@
 import Head from "next/head";
 import { useRouter } from "next/router";
 
-import Header from "@/components/header";
-import Info from "@/components/info";
-import Track from "@/components/track/track";
+import Index from "@/components/Index";
 
 const en = {
   title: "Check your product status",
@@ -13,7 +11,7 @@ const pt = {
   title: "Verifique o status do seu produto",
 };
 
-export default function TrackPage() {
+export default function IndexPage() {
   const router = useRouter();
   const t = router.locale === "en" ? en : pt;
 
@@ -23,12 +21,7 @@ export default function TrackPage() {
         <title>{t.title}</title>
       </Head>
 
-      <Header />
-
-      <main style={{ display: "contents" }}>
-        <Track />
-        <Info />
-      </main>
+      <Index />
     </>
   );
 }
